+++ CHANGELOG.md
@@ -13,4 +13,5 @@
 \- Multi-OS engine updated to version 1.4.1
 
 **[1.1.2]**
+\- GdxFIRLogger for debugging